Add files via upload
I created this using the Python programming language, with the following libraries:
import tkinter as tk
from tkinter import messagebox
import re
from datetime import datetime

to create a simple application with a graphical interface. By using tkinter, the application can display windows, buttons, and text as the user interface. Additionally, messagebox is used to show messages or alerts to the user. This code also utilizes re (Regular Expression) to search for or check patterns in text, as well as datetime to manage dates and times, such as recording when an event occurs. With this combination of libraries, the application can have a graphical interface, display notifications, process text, and handle time-related information efficiently.